+++ lib/mix/lib/mix/cli.ex
@@ -22,17 +22,14 @@ defmodule Mix.CLI do
     args = load_mixfile(args)
     { task, args } = get_task(args)
     change_env(task)
-
-    run_task task, args
+    run_task(task, args)
   end
 
   defp load_mixfile(args) do
     file = "mix.exs"
-
     if File.regular?(file) do
-      Code.load_file file
+      Code.load_file(file)
     end
-
     args
   end
 
@@ -52,7 +49,11 @@ defmodule Mix.CLI do
 
   defp run_task(name, args) do
     try do
-      if Mix.Project.get do
+      # We need to skip loading the paths for the project and
+      # its dependencies because the dependencies may be setting
+      # up or getting updated and we don't want to load their
+      # config until the process is done.
+      if not deps_task?(name) && Mix.Project.get do
         Mix.Task.run "loadpaths", ["--no-deps-check", "--no-elixir-version-check"]
         Mix.Task.reenable "loadpaths"
         Mix.Task.reenable "deps.loadpaths"
@@ -77,7 +78,8 @@ defmodule Mix.CLI do
   defp change_env(task) do
     if nil?(System.get_env("MIX_ENV")) && (env = Mix.project[:preferred_cli_env][task]) do
       Mix.env(env)
-      Mix.Project.push Mix.Project.pop
+      { project, file } = Mix.Project.pop
+      Mix.Project.push project, file
     end
   end
 
@@ -88,6 +90,10 @@ defmodule Mix.CLI do
   defp display_version() do
     IO.puts "Elixir #{System.version}"
   end
+
+  defp deps_task?("deps.update"), do: true
+  defp deps_task?("deps.get"), do: true
+  defp deps_task?(_), do: false
 
   # Check for --help or --version in the args
   defp check_for_shortcuts([first_arg|_]) when first_arg in