+++ lib/mix/test/mix/deps_test.exs
@@ -11,7 +11,8 @@ defmodule Mix.DepsTest do
           { :invalidvsn, "0.2.0", path: "deps/invalidvsn" },
           { :invalidapp, "0.1.0", path: "deps/invalidapp" },
           { :noappfile,  "0.1.0", path: "deps/noappfile" },
-          { :uncloned,            git: "https://github.com/elixir-lang/uncloned.git" }
+          { :uncloned,            git: "https://github.com/elixir-lang/uncloned.git" },
+          { :optional,            git: "https://github.com/elixir-lang/optional.git", optional: true }
         ]
       ]
     end
@@ -33,12 +34,14 @@ defmodule Mix.DepsTest do
     Mix.Project.push DepsApp
 
     in_fixture "deps_status", fn ->
-      deps = Mix.Deps.all
+      deps = Mix.Deps.fetched
+      assert length(deps) == 6
       assert Enum.find deps, &match?(Mix.Dep[app: :ok, status: { :ok, _ }], &1)
       assert Enum.find deps, &match?(Mix.Dep[app: :invalidvsn, status: { :invalidvsn, :ok }], &1)
       assert Enum.find deps, &match?(Mix.Dep[app: :invalidapp, status: { :invalidapp, _ }], &1)
       assert Enum.find deps, &match?(Mix.Dep[app: :noappfile, status: { :noappfile, _ }], &1)
       assert Enum.find deps, &match?(Mix.Dep[app: :uncloned, status: { :unavailable, _ }], &1)
+      assert Enum.find deps, &match?(Mix.Dep[app: :optional, status: { :unavailable, _ }], &1)
     end
   after
     Mix.Project.pop
@@ -48,7 +51,7 @@ defmodule Mix.DepsTest do
     Mix.Project.push MixVersionApp
 
     in_fixture "deps_status", fn ->
-      deps = Mix.Deps.all
+      deps = Mix.Deps.fetched
       assert Enum.find deps, &match?(Mix.Dep[app: :ok, status: { :ok, _ }], &1)
     end
   after
@@ -61,7 +64,70 @@ defmodule Mix.DepsTest do
     in_fixture "deps_status", fn ->
       msg = "Mix.DepsTest.NoSCMApp did not specify a supported scm for app :ok, " <>
             "expected one of :git, :path or :in_umbrella"
-      assert_raise Mix.Error, msg, fn -> Mix.Deps.all end
+      assert_raise Mix.Error, msg, fn -> Mix.Deps.fetched end
+    end
+  after
+    Mix.Project.pop
+  end
+
+  test "does not set the manager before the dependency was fetched" do
+    # It is important to not eagerly set the manager because the dependency
+    # needs to be fetched (i.e. available in the filesystem) in order to get
+    # the proper manager.
+    Mix.Project.push DepsApp
+
+    { _, true } =
+      Mix.Deps.unfetched(false, fn dep, acc ->
+        assert nil?(dep.manager)
+        { dep, acc or true }
+      end)
+  after
+    Mix.Project.pop
+  end
+
+  defmodule NestedDepsApp do
+    def project do
+      [
+        app: :raw_sample,
+        version: "0.1.0",
+        deps: [
+          { :deps_repo, "0.1.0", path: "custom/deps_repo" }
+        ]
+      ]
+    end
+  end
+
+  test "nested deps come first" do
+    Mix.Project.push NestedDepsApp
+
+    in_fixture "deps_status", fn ->
+      assert Enum.map(Mix.Deps.fetched, &(&1.app)) == [:git_repo, :deps_repo]
+    end
+  after
+    Mix.Project.pop
+  end
+
+  test "nested optional deps are never added" do
+    Mix.Project.push NestedDepsApp
+
+    in_fixture "deps_status", fn ->
+      File.write!("custom/deps_repo/mix.exs", """)
+      defmodule DepsRepo do
+        use Mix.Project
+
+        def project do
+          [
+            app: :deps_repo,
+            version: "0.1.0",
+            deps: [
+              { :git_repo, "0.2.0", git: MixTest.Case.fixture_path("git_repo"), optional: true }
+            ]
+          ]
+        end
+      end
+      """
+
+      assert Enum.map(Mix.Deps.fetched, &(&1.app)) == [:deps_repo]
     end
   after
     Mix.Project.pop
@@ -80,11 +146,37 @@ defmodule Mix.DepsTest do
     end
   end
 
-  test "correctly order overriden deps" do
+  test "correctly order converged deps" do
     Mix.Project.push ConvergedDepsApp
 
     in_fixture "deps_status", fn ->
-      assert [:git_repo, :deps_repo] == Enum.map(Mix.Deps.all, &(&1.app))
+      assert Enum.map(Mix.Deps.fetched, &(&1.app)) == [:git_repo, :deps_repo]
+    end
+  after
+    Mix.Project.pop
+  end
+
+  test "correctly order converged deps even with optional dependencies" do
+    Mix.Project.push ConvergedDepsApp
+
+    in_fixture "deps_status", fn ->
+      File.write!("custom/deps_repo/mix.exs", """)
+      defmodule DepsRepo do
+        use Mix.Project
+
+        def project do
+          [
+            app: :deps_repo,
+            version: "0.1.0",
+            deps: [
+              { :git_repo, "0.2.0", git: MixTest.Case.fixture_path("git_repo"), optional: true }
+            ]
+          ]
+        end
+      end
+      """
+
+      assert Enum.map(Mix.Deps.fetched, &(&1.app)) == [:git_repo, :deps_repo]
     end
   after
     Mix.Project.pop